+++ etc/afpd/fork.c
@@ -1281,7 +1281,7 @@ static int write_fork(AFPObj *obj, char *ibuf, size_t ibuflen _U_, char *rbuf, s
 
 #ifdef DEBUG1
         if (obj->options.flags & OPTION_DEBUG) {
-            printf("(write) len: %ld\n", *rbuflen);
+            printf("(write) len: %ld\n", (unsigned long) *rbuflen);
             bprint(rbuf, *rbuflen);
         }
 #endif
